inversiones-bursatiles/requirements.txt
yfinance==0.2.54
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
requests==2.32.3
python-telegram-bot==21.5
apscheduler==3.10.4
jinja2==3.1.4
python-dotenv==1.0.1
pytz==2024.1
aiohttp==3.9.5
httpx==0.27.0
